+++ app/upgrades/v5/constants.go
@@ -29,7 +29,7 @@ const (
 )
 
 var (
-	// MainnetMinGasPrices defines 20B aevmos (or atevmos) as the minimum gas price value on the fee market module.
+	// MainnetMinGasPrices defines 20B uCC (or atevmos) as the minimum gas price value on the fee market module.
 	// See https://commonwealth.im/evmos/discussion/5073-global-min-gas-price-value-for-cosmos-sdk-and-evm-transaction-choosing-a-value for reference
 	MainnetMinGasPrices = sdk.NewDec(20_000_000_000)
 	// MainnetMinGasMultiplier defines the min gas multiplier value on the fee market module.