+++ x/claims/types/params_test.go
@@ -128,7 +128,7 @@ func TestParamsvalidateDuration(t *testing.T) {
 }
 
 func TestParamsValidateDenom(t *testing.T) {
-	err := validateDenom("aevmos")
+	err := validateDenom("uCC")
 	require.NoError(t, err)
 	err = validateDenom(false)
 	require.Error(t, err)