+++ x/incentives/keeper/distribution.go
@@ -217,7 +217,7 @@ func (k Keeper) rewardParticipants(
 					continue
 				}
 
-				// Cap rewards in mint denom (i.e. aevmos) to receive only up to 100% of
+				// Cap rewards in mint denom (i.e. uCC) to receive only up to 100% of
 				// the participant's gas spent and prevent gaming
 				if mintDenom == allocation.Denom {
 					rewardCap := cumulativeGas.Mul(rewardScaler)